requirements_old.txt
six
oauthlib>=3.0.0
markdown>=2.6.8
pillow!=8.3.*,>=5.3.0
appdirs>=1.4.3
Click!=8.0.0,>=7.1
psutil>=5.0.0
idna
charset-normalizer
attrs
requests
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
black==23.12.1
bumpversion==0.6.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.2.0
datasets==2.16.1
dill==0.3.7
docker-pycreds==0.4.0
editdistance==0.6.2
efficientnet-pytorch==0.7.1
filelock==3.13.1
flatbuffers==23.5.26
fpdf==1.7.2
frozenlist==1.4.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud==0.34.0
google-cloud-documentai==2.24.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
huggingface-hub==0.19.4
hydra-core==1.3.2
idna==3.6
imageio==2.32.0
imgaug==0.2.6
install==1.3.5
iopath==0.1.9
ipywidgets==8.1.1
joblib==1.3.2
jupyterlab-widgets==3.0.9
keras==2.15.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.10.0
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opensearch-py==2.2.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pathspec==0.12.1
pdf2image==1.17.0
portalocker==2.8.2
pretrainedmodels==0.7.4
proto-plus==1.23.0
protobuf==4.23.4
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyDeprecate==0.3.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-lightning==2.0.0
PyYAML==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
segmentation-models-pytorch==0.3.3
sentencepiece==0.1.99
sentry-sdk==1.40.5
setproctitle==1.3.3
shapely==2.0.2
smmap==5.0.1
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
tokenizers==0.15.0
tomli==2.0.1
torch==1.11.0+cu115
torchmetrics==1.2.1
torchvision==0.12.0+cu115
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wandb==0.16.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
